requirements.txt
streamlit
rdkit-pypi